unify UIView Extensions
